+++ RELEASENOTES.md
@@ -13,3 +13,5 @@
 - Fix touchy to not work with ripper projectiles.
 - Fix A_SpawnObject to clear velocity if object is spawned below floor or above ceiling.
 - Fix missiles not exploding on floors when floor hits the missile.
+- Fix MF_BOUNCES flag to account for mass and implement velocity modifications from Boom.
+- Fix dropoff check to work correctly MF_BOUNCES things and things on midtex3d lines crossing sectors.